Move Dashboard.tools functions to utils
